alarm API : 복합키 직전 백업

+++ backend/src/main/java/com/bizkicks/backend/service/AlarmService.java
@@ -1,7 +1,10 @@
 package com.bizkicks.backend.service;
 
 import com.bizkicks.backend.entity.Alarm;
+import com.bizkicks.backend.exception.CustomException;
+import com.bizkicks.backend.exception.ErrorCode;
 import com.bizkicks.backend.repository.AlarmRepository;
+
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,11 +21,13 @@ public class AlarmService {
 
     @Transactional
     public List<Alarm> findAlarms(String customerCompanyName){
+        if(!alarmRepository.isCustomerCompanyExist(customerCompanyName)) throw new CustomException(ErrorCode.COMPANY_NOT_EXIST);
         return alarmRepository.findByCustomerCompanyName(customerCompanyName);
     }
 
     @Transactional
     public void updateAlarms(String customerCompanyName, List<Alarm> alarms){
+        if(!alarmRepository.isCustomerCompanyExist(customerCompanyName)) throw new CustomException(ErrorCode.COMPANY_NOT_EXIST);
         alarmRepository.deleteAllAlarmsInCustomerCompany(customerCompanyName);
         alarmRepository.saveAllAlarmsInCustomerCompany(customerCompanyName, alarms);
     }
